module file is renamed to fileSync, SourceManager class is renamed to FileSync
