+++ org.eclipse.buildship.kts.scripts/language-server-for-gradle-properties/app/src/main/java/server/fileSync/ContentInFile.java
@@ -1,4 +1,4 @@
-package server.file;
+package server.fileSync;
 
 public class ContentInFile {
